+++ NEWS.md
@@ -33,6 +33,9 @@ Breaking changes
 
 This section lists changes that do not have deprecation warnings.
 
+  * `readline`, `readlines` and `eachline` return lines without line ends by default.
+     You can use `readline(s, false`) to get the old behavior and include EOL character(s). ([#19944]).
+
   * `String`s no longer have a `.data` field (as part of a significant performance
     improvement). Use `Vector{UInt8}(str)` to access a string as a byte array.
     However, allocating the `Vector` object has overhead. You can also use
@@ -164,9 +167,6 @@ Compiler/Runtime improvements
 
 Deprecated or removed
 ---------------------
-
-  * One argument methods to `readline`, `readlines` and `eachline` have been deprecated in
-    order to add new `chomp::Bool` argument which will default to `true` in future versions ([#19944]).
 
   * `isdefined(a::Array, i::Int)` has been deprecated in favor of `isassigned` ([#18346]).